requirements.txt
requests
beautifulsoup4
python-telegram-bot
schedule
pandas
sqlalchemy
lxml
python-dotenv
